restaurant-assistant/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
pydantic
python-dotenv
sentence-transformers
chromadb
email-validator
